requirements.txt
# 2026 Stable Requirements for Restaurant Generator
streamlit>=1.40.0
langchain>=1.2.0
langchain-google-genai>=4.2.0
langchain-core>=1.2.0
python-dotenv>=1.0.1
google-genai>=1.59.0